# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Business Analytics Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DeleteResult {
  success: boolean;
}

export interface ChartDataPoint {
  label: string;
  value: number;
  /** @nullable */
  secondary?: number | null;
}

export interface CategoryStat {
  name: string;
  value: number;
  percentage: number;
  /** @nullable */
  color?: string | null;
}

export interface RevenueProfitPoint {
  label: string;
  revenue: number;
  profit: number;
}

export interface TopProduct {
  id: number;
  name: string;
  unitsSold: number;
  revenue: number;
  /** @nullable */
  rank?: number | null;
}

export interface AiInsight {
  type: string;
  title: string;
  message: string;
}

export interface DashboardStats {
  totalRevenue: number;
  totalOrders: number;
  totalProfit: number;
  totalProducts: number;
  newCustomers: number;
  revenueChange?: number;
  ordersChange?: number;
  profitChange?: number;
  productsChange?: number;
  customersChange?: number;
}

export interface Order {
  id: number;
  orderId: string;
  product: string;
  amount: number;
  status: string;
  date: string;
}

export interface InventoryStatusItem {
  id: number;
  product: string;
  stock: number;
  status: string;
}

export interface Sale {
  id: number;
  invoiceId: string;
  date: string;
  customer: string;
  product: string;
  category: string;
  qty: number;
  unitPrice: number;
  revenue: number;
  channel: string;
}

export interface SaleInput {
  customer: string;
  product: string;
  category: string;
  qty: number;
  unitPrice: number;
  channel: string;
  /** @nullable */
  date?: string | null;
}

export interface SaleUpdate {
  customer?: string;
  product?: string;
  category?: string;
  qty?: number;
  unitPrice?: number;
  channel?: string;
}

export interface SalesList {
  data: Sale[];
  total: number;
  page: number;
  limit: number;
}

export interface SalesStats {
  totalRevenue: number;
  totalUnitsSold: number;
  totalTransactions: number;
  newCustomers: number;
  revenueChange?: number;
  unitsSoldChange?: number;
  transactionsChange?: number;
  newCustomersChange?: number;
}

export interface InventoryItem {
  id: number;
  sku: string;
  product: string;
  category: string;
  warehouse: string;
  stock: number;
  threshold: number;
  status: string;
  stockValue: number;
}

export interface InventoryItemInput {
  sku: string;
  product: string;
  category: string;
  warehouse: string;
  stock: number;
  threshold: number;
  /** @nullable */
  unitCost?: number | null;
}

export interface InventoryItemUpdate {
  product?: string;
  category?: string;
  warehouse?: string;
  stock?: number;
  threshold?: number;
  /** @nullable */
  unitCost?: number | null;
}

export interface InventoryList {
  data: InventoryItem[];
  total: number;
  page: number;
  limit: number;
}

export interface InventoryStats {
  totalProducts: number;
  totalStock: number;
  lowStockItems: number;
  outOfStockItems: number;
  inventoryValue: number;
}

export interface StockStatusItem {
  status: string;
  count: number;
  percentage: number;
}

export interface LowStockAlert {
  id: number;
  product: string;
  stock: number;
  threshold: number;
}

export interface StockMovement {
  id: number;
  product: string;
  warehouse: string;
  change: number;
  date: string;
}

export interface ForecastChartPoint {
  label: string;
  /** @nullable */
  historical: number | null;
  /** @nullable */
  forecasted: number | null;
  /** @nullable */
  upperBound?: number | null;
  /** @nullable */
  lowerBound?: number | null;
}

export interface ForecastResult {
  forecastedSales: number;
  avgDailyDemand: number;
  peakDemandDay: string;
  peakDemandUnits?: number;
  totalRevenueForecast: number;
  recommendedStock: number;
  mape: number;
  rmse: number;
  accuracy: number;
  chartData: ForecastChartPoint[];
}

export interface ForecastBreakdownRow {
  period: string;
  forecastedDemand: number;
  lowerBound: number;
  upperBound: number;
  revenueForecast: number;
}

export interface HeatmapRow {
  week: number;
  mon: number;
  tue: number;
  wed: number;
  thu: number;
  fri: number;
  sat: number;
  sun: number;
}

export interface ModelComparison {
  model: string;
  mape: number;
  isRecommended: boolean;
}

export interface ReportRow {
  date: string;
  product: string;
  category: string;
  orders: number;
  unitsSold: number;
  revenue: number;
  profit: number;
}

export interface ReportData {
  data: ReportRow[];
  total: number;
  page: number;
  limit: number;
}

export interface ReportStats {
  totalRevenue: number;
  totalOrders: number;
  totalUnitsSold: number;
  avgOrderValue: number;
  revenueChange?: number;
  ordersChange?: number;
  unitsSoldChange?: number;
}

export interface Customer {
  id: number;
  name: string;
  email: string;
  phone: string;
  segment: string;
  totalOrders: number;
  totalSpent: number;
  joinDate: string;
  /** @nullable */
  location?: string | null;
}

export interface CustomerInput {
  name: string;
  email: string;
  phone: string;
  segment: string;
  /** @nullable */
  location?: string | null;
}

export interface CustomerUpdate {
  name?: string;
  email?: string;
  phone?: string;
  segment?: string;
  /** @nullable */
  location?: string | null;
}

export interface CustomerList {
  data: Customer[];
  total: number;
  page: number;
  limit: number;
}

export interface CustomerStats {
  totalCustomers: number;
  newCustomers: number;
  totalOrders: number;
  totalRevenue: number;
  avgCustomerValue: number;
  totalCustomersChange?: number;
  newCustomersChange?: number;
  totalOrdersChange?: number;
  totalRevenueChange?: number;
}

export interface TopCustomer {
  id: number;
  name: string;
  revenue: number;
  segment: string;
}

export interface LocationStat {
  location: string;
  count: number;
  percentage: number;
}

export interface AppSettings {
  businessName: string;
  businessEmail: string;
  defaultTheme: string;
  language: string;
  timezone: string;
  currency: string;
  dateFormat: string;
  timeFormat: string;
  enableAnalytics: boolean;
  autoRefresh: boolean;
  emailNotifications: boolean;
  darkMode: boolean;
  compactView: boolean;
}

export interface SettingsUpdate {
  businessName?: string;
  businessEmail?: string;
  defaultTheme?: string;
  language?: string;
  timezone?: string;
  currency?: string;
  dateFormat?: string;
  timeFormat?: string;
  enableAnalytics?: boolean;
  autoRefresh?: boolean;
  emailNotifications?: boolean;
  darkMode?: boolean;
  compactView?: boolean;
}

export type GetDashboardRevenueOverviewParams = {
  period?: string;
};

export type GetDashboardRevenueVsProfitParams = {
  period?: string;
};

export type GetSalesParams = {
  page?: number;
  limit?: number;
  search?: string;
  category?: string;
  channel?: string;
  dateFrom?: string;
  dateTo?: string;
};

export type GetSalesTrendParams = {
  period?: string;
};

export type GetInventoryParams = {
  page?: number;
  limit?: number;
  search?: string;
  category?: string;
  warehouse?: string;
  stockStatus?: string;
};

export type GetForecastParams = {
  productId?: number;
  model?: string;
  horizon?: number;
};

export type GetForecastBreakdownParams = {
  productId?: number;
  model?: string;
  horizon?: number;
};

export type GetForecastHeatmapParams = {
  productId?: number;
};

export type GetReportParams = {
  reportType?: string;
  dateFrom?: string;
  dateTo?: string;
  category?: string;
  product?: string;
  page?: number;
  limit?: number;
};

export type GetReportStatsParams = {
  dateFrom?: string;
  dateTo?: string;
};

export type GetCustomersParams = {
  page?: number;
  limit?: number;
  search?: string;
  segment?: string;
  group?: string;
  location?: string;
};