# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Business Analytics Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: getInventory
      tags: [inventory]
      summary: List inventory items
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: warehouse
          in: query
          schema:
            type: string
        - name: stockStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Inventory list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryList"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Add a new product to inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
...
  /inventory/{id}:
    put:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update an inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemUpdate"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete an inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  # ─── FORECASTING ──────────────────────────────────────────────
...
  /customers:
    get:
      operationId: getCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          schema:
            type: string
        - name: segment
          in: query
          schema:
            type: string
        - name: group
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Customers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerList"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Add a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
...
  /customers/{id}:
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  # ─── SETTINGS ────────────────────────────────────────────────
...
components:
  schemas:
...
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    InventoryItem:
      type: object
      required: [id, sku, product, category, warehouse, stock, threshold, status, stockValue]
      properties:
        id:
          type: integer
        sku:
          type: string
        product:
          type: string
        category:
          type: string
        warehouse:
          type: string
        stock:
          type: integer
        threshold:
          type: integer
        status:
          type: string
        stockValue:
          type: number
    InventoryItemInput:
      type: object
      required: [sku, product, category, warehouse, stock, threshold]
      properties:
        sku:
          type: string
        product:
          type: string
        category:
          type: string
        warehouse:
          type: string
        stock:
          type: integer
        threshold:
          type: integer
        unitCost:
          type: ["number", "null"]
    InventoryItemUpdate:
      type: object
      properties:
        product:
          type: string
        category:
          type: string
        warehouse:
          type: string
        stock:
          type: integer
        threshold:
          type: integer
        unitCost:
          type: ["number", "null"]
    InventoryList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/InventoryItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    Customer:
      type: object
      required: [id, name, email, phone, segment, totalOrders, totalSpent, joinDate]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        segment:
          type: string
        totalOrders:
          type: integer
        totalSpent:
          type: number
        joinDate:
          type: string
        location:
          type: ["string", "null"]
    CustomerInput:
      type: object
      required: [name, email, phone, segment]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        segment:
          type: string
        location:
          type: ["string", "null"]
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        segment:
          type: string
        location:
          type: ["string", "null"]
    CustomerList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...