# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Business Analytics Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerUpdate {
  name?: string;
  email?: string;
  phone?: string;
  segment?: string;
  /** @nullable */
  location?: string | null;
}