# lib/api-zod/src/generated/types/inventoryItemUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Business Analytics Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryItemUpdate {
  product?: string;
  category?: string;
  warehouse?: string;
  stock?: number;
  threshold?: number;
  /** @nullable */
  unitCost?: number | null;
}